Refactor import Functions
Jira ticket: CAMS-421
