+++ backend/function-apps/dataflows/import/migrate-cases.ts
@@ -1,152 +1,64 @@
-import * as df from 'durable-functions';
-import { OrchestrationContext } from 'durable-functions';
-import { app, HttpRequest, HttpResponse, InvocationContext } from '@azure/functions';
+import { app, HttpRequest, HttpResponse, InvocationContext, output } from '@azure/functions';
 
-import {
-  CaseSyncEvent,
-  ExportCaseChangeEventsSummary,
-  getDefaultSummary,
-} from '../../../../common/src/queue/dataflow-types';
+import { CaseSyncEvent } from '../../../../common/src/queue/dataflow-types';
 import { ForbiddenError } from '../../../lib/common-errors/forbidden-error';
 
 import ContextCreator from '../../azure/application-context-creator';
-import { toAzureError } from '../../azure/functions';
+import { toAzureError, toAzureSuccess } from '../../azure/functions';
 import { buildUniqueName, isAuthorized } from '../dataflows-common';
-import { STORE_CASES_RUNTIME_STATE } from './store-cases-runtime-state';
 import MigrateCases from '../../../lib/use-cases/dataflows/migrate-cases';
-import { DLQ } from '../dataflows-queues';
-import { EXPORT_AND_LOAD_CASE } from './export-and-load-case';
 import { buildQueueError } from '../../../lib/use-cases/dataflows/queue-types';
+import { storeCasesRuntimeState } from './store-cases-runtime-state';
 
-const MODULE_NAME = 'MIGRATE_CASES_DATAFLOW';
+const MODULE_NAME = 'MIGRATE_CASES';
 
-// Orchestration Aliases
-const MIGRATE_CASES = buildUniqueName(MODULE_NAME, 'migrateCases');
-const PARTITION_CASEIDS = buildUniqueName(MODULE_NAME, 'partitionCaseIds');
-const MIGRATE_PARTITION = buildUniqueName(MODULE_NAME, 'migratePartition');
+const PAGE_SIZE = 1000;
+
+// Queues
+let START;
+let PAGE;
+let DLQ;
 
 // Activity Aliases
-const GET_CASEIDS_TO_MIGRATE_ACTIVITY = buildUniqueName(MODULE_NAME, 'getCaseIdsToMigrateActivity');
-const LOAD_MIGRATION_TABLE = buildUniqueName(MODULE_NAME, 'loadMigrationTable');
-const EMPTY_MIGRATION_TABLE = buildUniqueName(MODULE_NAME, 'emptyMigrationTable');
+const GET_CASEIDS_TO_MIGRATE_ACTIVITY = buildUniqueName(MODULE_NAME, 'getCaseIdsToMigrate');
+const LOAD_MIGRATION_TABLE_ACTIVITY = buildUniqueName(MODULE_NAME, 'loadMigrationTable');
+const EMPTY_MIGRATION_TABLE_ACTIVITY = buildUniqueName(MODULE_NAME, 'emptyMigrationTable');
+const STORE_CASES_RUNTIME_STATE = buildUniqueName(MODULE_NAME, 'storeCasesRuntimeState');
 
 /**
  * loadMigrationTable
  *
- * @param _ignore
  * @param invocationContext
  * @returns
  */
-async function loadMigrationTable(_ignore: unknown, invocationContext: InvocationContext) {
+async function loadMigrationTable(invocationContext: InvocationContext) {
   const context = await ContextCreator.getApplicationContext({ invocationContext });
   const result = await MigrateCases.loadMigrationTable(context);
   if (result.error) {
     invocationContext.extraOutputs.set(
       DLQ,
-      buildQueueError(result.error, MODULE_NAME, LOAD_MIGRATION_TABLE),
+      buildQueueError(result.error, MODULE_NAME, LOAD_MIGRATION_TABLE_ACTIVITY),
     );
   }
   return result.data;
 }
 
 /**
- * partitionCaseIds
- *
- * @param context
- */
-function* partitionCaseIds(context: OrchestrationContext) {
-  const count: number = context.df.getInput();
-
-  const finalSummary = getDefaultSummary();
-  const partitionSize = 1000;
-
-  let start = 0;
-  let end = 0;
-  let partitionCount = 0;
-
-  while (end < count) {
-    partitionCount += 1;
-    start = end + 1;
-    end += partitionSize;
-    const task = yield context.df.callSubOrchestrator(
-      MIGRATE_PARTITION,
-      { start, end },
-      context.df.instanceId + `:${MIGRATE_CASES}:partition:${partitionCount}`,
-    );
-
-    if (task.result) {
-      const result = task.result as ExportCaseChangeEventsSummary;
-      finalSummary.changedCases += result.changedCases;
-      finalSummary.exportedAndLoaded += result.exportedAndLoaded;
-      finalSummary.completed += result.completed;
-      finalSummary.errors += result.errors;
-      finalSummary.faulted += result.faulted;
-      finalSummary.noResult += result.noResult;
-    }
-  }
-
-  return finalSummary;
-}
-
-function* migratePartition(context: OrchestrationContext) {
-  const range = context.df.getInput();
-
-  const events: CaseSyncEvent[] = yield context.df.callActivity(
-    GET_CASEIDS_TO_MIGRATE_ACTIVITY,
-    range,
-  );
-
-  const nextTasks: df.Task[] = [];
-
-  for (const event of events) {
-    const childId = context.df.instanceId + `:${EXPORT_AND_LOAD_CASE}:${event.caseId}:`;
-    nextTasks.push(context.df.callSubOrchestrator(EXPORT_AND_LOAD_CASE, event, childId));
-  }
-
-  yield context.df.Task.all(nextTasks);
-
-  const results = nextTasks.reduce(
-    (summary, task) => {
-      if (task.isCompleted) {
-        summary.completed += 1;
-      }
-      if (task.isFaulted) {
-        summary.faulted += 1;
-      }
-      if (task.result) {
-        const event = task.result as unknown as CaseSyncEvent;
-        if (event.error) {
-          summary.errors += 1;
-        } else {
-          summary.exportedAndLoaded += 1;
-        }
-      } else {
-        summary.noResult += 1;
-      }
-
-      return summary;
-    },
-    getDefaultSummary({ changedCases: events.length }),
-  );
-
-  return results;
-}
-
-/**
  * emptyMigrationTable
  *
- * @param _ignore
  * @param invocationContext
  */
-async function emptyMigrationTable(_ignore: unknown, invocationContext: InvocationContext) {
+async function emptyMigrationTable(invocationContext: InvocationContext) {
   const context = await ContextCreator.getApplicationContext({ invocationContext });
   const result = await MigrateCases.emptyMigrationTable(context);
   if (result.error) {
     invocationContext.extraOutputs.set(
       DLQ,
-      buildQueueError(result.error, MODULE_NAME, EMPTY_MIGRATION_TABLE),
+      buildQueueError(result.error, MODULE_NAME, EMPTY_MIGRATION_TABLE_ACTIVITY),
     );
+    return false;
   }
+  return true;
 }
 
 /**
@@ -180,41 +92,69 @@ async function getCaseIdsToMigrate(
 }
 
 /**
+ * processPage
+ *
+ * Get case Ids from ACMS identifying cases to migrate then export and load the cases from DXTR into CAMS.
+ *
+ * @param range
+ * @param invocationContext
+ */
+async function processPage(
+  range: { start: number; end: number },
+  invocationContext: InvocationContext,
+) {
+  const events: CaseSyncEvent[] = await getCaseIdsToMigrate(range, invocationContext);
+
+  const appContext = await ContextCreator.getApplicationContext({ invocationContext });
+  const processedEvents = await MigrateCases.exportAndLoadPage(appContext, events);
+
+  const failedEvents = processedEvents.filter((event) => !!event.error);
+  invocationContext.extraOutputs.set(DLQ, failedEvents);
+}
+
+/**
  * migrateCases
  *
  * Get case Ids from ACMS identifying cases to migrate then export and load the cases from DXTR into CAMS.
  *
- * @param  context
+ * @param {object} message
+ * @param {InvocationContext} context
  */
-function* migrateCases(context: OrchestrationContext) {
-  yield context.df.callActivity(EMPTY_MIGRATION_TABLE);
-  const count = yield context.df.callActivity(LOAD_MIGRATION_TABLE);
+async function migrateCases(message: { invocationId: string }, context: InvocationContext) {
+  if (!message.invocationId) return;
 
-  if (count === 0) {
-    return getDefaultSummary({ changedCases: count });
+  const logger = ContextCreator.getLogger(context);
+  logger.info(MODULE_NAME, `Migrating cases. Invocation id: ${message.invocationId}.`);
+
+  const isEmpty = await emptyMigrationTable(context);
+  if (!isEmpty) return;
+
+  const count = await loadMigrationTable(context);
+
+  if (count === 0) return;
+
+  let start = 0;
+  let end = 0;
+
+  const pages = [];
+  while (end < count) {
+    start = end + 1;
+    end += PAGE_SIZE;
+    pages.push({ start, end });
   }
+  context.extraOutputs.set(PAGE, pages);
 
-  const childId = context.df.instanceId + `:${PARTITION_CASEIDS}`;
-  const summary = yield context.df.callSubOrchestrator(PARTITION_CASEIDS, count, childId);
-
-  yield context.df.callSubOrchestrator(
-    STORE_CASES_RUNTIME_STATE,
-    {},
-    context.df.instanceId + `:${MIGRATE_CASES}:${STORE_CASES_RUNTIME_STATE}`,
-  );
-  yield context.df.callActivity(EMPTY_MIGRATION_TABLE);
-
-  return summary;
+  await storeCasesRuntimeState({ activityName: STORE_CASES_RUNTIME_STATE }, context);
 }
 
 /**
- * migrateCasesHttpTrigger
+ * httpTrigger
  *
  * @param request
  * @param context
  * @returns
  */
-async function migrateCasesHttpTrigger(
+async function httpTrigger(
   request: HttpRequest,
   context: InvocationContext,
 ): Promise<HttpResponse> {
@@ -222,34 +162,48 @@ async function migrateCasesHttpTrigger(
     if (!isAuthorized(request)) {
       throw new ForbiddenError(MODULE_NAME);
     }
-    const client = df.getClient(context);
-    const instanceId: string = await client.startNew(MIGRATE_CASES);
+    context.extraOutputs.set(START, { invocationId: context.invocationId });
 
-    return client.createCheckStatusResponse(request, instanceId);
+    return new HttpResponse(toAzureSuccess({ statusCode: 201 }));
   } catch (error) {
     return new HttpResponse(toAzureError(ContextCreator.getLogger(context), MODULE_NAME, error));
   }
 }
 
 export function setupMigrateCases() {
-  df.app.orchestration(MIGRATE_CASES, migrateCases);
-  df.app.orchestration(PARTITION_CASEIDS, partitionCaseIds);
-  df.app.orchestration(MIGRATE_PARTITION, migratePartition);
-
-  df.app.activity(GET_CASEIDS_TO_MIGRATE_ACTIVITY, {
-    handler: getCaseIdsToMigrate,
+  START = output.storageQueue({
+    queueName: buildUniqueName(MODULE_NAME, 'start').toLowerCase(),
+    connection: 'AzureWebJobsStorage',
   });
 
-  df.app.activity(LOAD_MIGRATION_TABLE, {
-    handler: loadMigrationTable,
+  PAGE = output.storageQueue({
+    queueName: buildUniqueName(MODULE_NAME, 'page').toLowerCase(),
+    connection: 'AzureWebJobsStorage',
+  });
+
+  DLQ = output.storageQueue({
+    queueName: buildUniqueName(MODULE_NAME, 'dlq').toLowerCase(),
+    connection: 'AzureWebJobsStorage',
+  });
+
+  app.storageQueue(buildUniqueName(MODULE_NAME, 'migrateCases'), {
+    queueName: START.queueName,
+    connection: 'AzureWebJobsStorage',
+    handler: migrateCases,
+    extraOutputs: [PAGE],
+  });
+
+  app.storageQueue(buildUniqueName(MODULE_NAME, 'processPage'), {
+    queueName: PAGE.queueName,
+    connection: 'AzureWebJobsStorage',
+    handler: processPage,
     extraOutputs: [DLQ],
   });
 
-  df.app.activity(EMPTY_MIGRATION_TABLE, { handler: emptyMigrationTable, extraOutputs: [DLQ] });
-
-  app.http('migrateCasesHttpTrigger', {
+  app.http(buildUniqueName(MODULE_NAME, 'httpTrigger'), {
+    methods: ['GET'],
     route: 'migratecases',
-    extraInputs: [df.input.durableClient()],
-    handler: migrateCasesHttpTrigger,
+    extraOutputs: [START],
+    handler: httpTrigger,
   });
 }
